Department added

+++ JDBC_learning/src/main/java/model/entities/Department.java
@@ -5,22 +5,22 @@ import java.util.Objects;
 
 public class Department implements Serializable {
 
-    private Integer id;
+    private long id;
     private String name;
 
     public Department() {
     }
 
-    public Department(Integer id, String name) {
+    public Department(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
